% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fish.R
\docType{data}
\name{fishData}
\alias{fishData}
\title{A data set that provides useful predictor variables in order to predict
lifespan.}
\format{
A data frame with 1976 observations with 7 columns:
\describe{
\item{Average Length}{Average length of the fish in inches.}
\item{Average Weight}{Average weight of the fish in ounces}
\item{Habitat}{The type of habitat the fish lives in}
\item{PH of Water}{Level of PH in water}
\item{Color}{Color of the fish}
\item{Gender}{Gender of fish (Male/Female)}
\item{Lifespan}{The total lifespan of the fish}

}
}
\source{
\url{https://www.kaggle.com/datasets/stealthtechnologies/predict-lifespan-of-a-comet-goldfish}?
}
\usage{
fishData
}
\description{
This data was collected from Kaggle. The data covers 1976 goldfish and the
common predictor variables of the goldfish. The source for the data can be
found below in the URL link attatched.
}
\keyword{datasets}
